Create modules for authors and games methods
