+++ app.rb
@@ -1,11 +1,11 @@
-require './game'
-require './author'
-
-['./book'].each do |file|
+['./book', './game', './author', './game_methods', './author_methods'].each do |file|
   require file
 end
 
 class App
+  include GameMethods
+  include AuthorMethods
+
   def initialize
     @books = []
     @labels = []
@@ -47,83 +47,18 @@ class App
     @books << Book.new(publisher, cover_state, publish_date)
   end
 
-  def list_games
-    return puts 'There are no games yet' if @games.empty?
-
-    puts 'List of games:'
-    @games.each_with_index do |game, index|
-      puts "#{index + 1}) #{game.multiplayer} #{game.last_played_at}, id=#{game.id}"
-    end
-  end
-
-  def list_authors
-    return puts 'There are no authors yet' if @authors.empty?
-
-    puts 'List of authors:'
-    @authors.each_with_index do |author, index|
-      puts "#{index + 1}) #{author.first_name} #{author.last_name}, id=#{author.id}"
-    end
-  end
-
-  def create_author
-    puts 'Enter first name:'
-    print '> '
-    first_name = gets.chomp
-    puts 'Enter last name:'
-    print '> '
-    last_name = gets.chomp
-    puts 'Author created successfully!'
-
-    @authors << Author.new(first_name, last_name)
-  end
-
-  def list_authors_to_select
-    puts '0) Create a new author'
-    list_authors
-    print '> '
-    selection = gets.chomp
-
-    if selection == '0'
-      create_author
-      return list_authors_to_select
-    end
-
-    return print "Invalid option, try again.\n\n" if invalid?(selection, @authors.length)
-
-    @authors[selection.to_i - 1]
-  end
-
   def invalid?(option, max_value)
     return true if option.to_i > max_value || /\D/.match?(option)
 
     false
   end
 
-  def game_details
-    puts 'Please enter multiplayer mode:'
-    print '> '
-    multiplayer = gets.chomp
-    puts 'Please enter the last date when the game was played (yyyy-mm-dd)'
-    print '> '
-    last_played_at = gets.chomp
-    puts 'Please provide publish date (yyyy-mm-dd)'
-    print '> '
-    publish_date = gets.chomp
-    puts 'Select an author from the list or create a new one'
-    author = list_authors_to_select
-    [multiplayer, last_played_at, publish_date, author]
+  def valid_date?(date)
+    date_format = '%Y-%m-%d'
+    DateTime.strptime(date, date_format)
+    true
+  rescue ArgumentError
+    false
   end
 
-  def add_game
-    multiplayer, last_played_at, publish_date, author = game_details
-
-    return if author.nil?
-
-    new_game = Game.new(multiplayer, last_played_at, publish_date)
-    new_game.add_author(author)
-
-    @games << new_game
-
-    puts 'Game created sucessfully!'
-  end
 end
