+++ game.rb
@@ -5,7 +5,7 @@ class Game < Item
   attr_reader :multiplayer, :last_played_at
 
   def initialize(multiplayer, last_played_at, publish_date, id = rand(1..1000), archived: false)
-    super(publish_date, id, archived)
+    super(publish_date, id, archived: archived)
     @multiplayer = multiplayer
     @last_played_at = Date.strptime(last_played_at, '%Y-%m-%d')
   end
